+++ NEWS.md
@@ -1,4 +1,5 @@
 ### crunch 1.10.1 (under development)
+* Add `icon` and `icon<-` methods for Projects to read the project's current icon URL and to set a new icon by supplying a local file name to upload.
 
 ## crunch 1.10.0